requirements.txt
python-telegram-bot==22.1
Pillow==10.4.0
qrcode[pil]==7.4.2